--- a/turbopack/crates/turbo-tasks-macros/src/func.rs
+++ b/turbopack/crates/turbo-tasks-macros/src/func.rs
@@ -480,17 +480,58 @@ impl TurboFn<'_> {
     }
 
     fn inline_input_idents(&self) -> impl Iterator<Item = &Ident> {
-        self.exposed_inputs
-            .iter()
-            .map(|Input { ident, .. }| ident)
+        self.exposed_input_idents()
             .filter(|id| inline_inputs_identifier_filter(id))
     }
 
-    pub fn input_types(&self) -> Vec<&Type> {
+    fn exposed_input_idents(&self) -> impl Iterator<Item = &Ident> {
+        self.exposed_inputs.iter().map(|Input { ident, .. }| ident)
+    }
+
+    pub fn exposed_input_types(&self) -> impl Iterator<Item = Cow<'_, Type>> {
         self.exposed_inputs
             .iter()
-            .map(|Input { ty, .. }| ty)
-            .collect()
+            .map(|Input { ty, .. }| expand_task_input_type(ty))
+    }
+
+    pub fn filter_trait_call_args(&self) -> Option<FilterTraitCallArgsTokens> {
+        // we only need to do this on trait methods, but we're doing it on all methods because we
+        // don't know if we're a trait method or not (we could pass this information down)
+        if self.is_method() {
+            let inline_input_idents: Vec<_> = self.inline_input_idents().collect();
+            if inline_input_idents.len() != self.exposed_inputs.len() {
+                let exposed_input_idents: Vec<_> = self.exposed_input_idents().collect();
+                let exposed_input_types: Vec<_> = self.exposed_input_types().collect();
+                return Some(FilterTraitCallArgsTokens {
+                    filter_owned: quote! {
+                        |magic_any| {
+                            let (#(#exposed_input_idents,)*) =
+                                *turbo_tasks::macro_helpers
+                                    ::downcast_args_owned::<(#(#exposed_input_types,)*)>(magic_any);
+                            ::std::boxed::Box::new((#(#inline_input_idents,)*))
+                        }
+                    },
+                    filter_and_resolve: quote! {
+                        |magic_any| {
+                            Box::pin(async move {
+                                let (#(#exposed_input_idents,)*) = turbo_tasks::macro_helpers
+                                    ::downcast_args_ref::<(#(#exposed_input_types,)*)>(magic_any);
+                                let resolved = (#(
+                                    <_ as turbo_tasks::TaskInput>::resolve(
+                                        #inline_input_idents
+                                    ).await?,
+                                )*);
+                                Ok(
+                                    ::std::boxed::Box::new(resolved)
+                                    as ::std::boxed::Box<dyn turbo_tasks::MagicAny>
+                                )
+                            })
+                        }
+                    },
+                });
+            }
+        }
+        None
     }
 
     pub fn persistence(&self) -> impl ToTokens {
@@ -572,7 +613,7 @@ impl TurboFn<'_> {
         let ident = &self.ident;
         let output = &self.output;
         let assertions = self.get_assertions();
-        let inputs = self.inline_input_idents();
+        let inputs = self.exposed_input_idents();
         let persistence = self.persistence_with_this();
         parse_quote! {
             {
@@ -1047,17 +1088,24 @@ impl DefinitionContext {
 }
 
 #[derive(Debug)]
+pub struct FilterTraitCallArgsTokens {
+    filter_owned: TokenStream,
+    filter_and_resolve: TokenStream,
+}
+
+#[derive(Debug)]
 pub struct NativeFn {
     pub function_path_string: String,
     pub function_path: ExprPath,
     pub is_method: bool,
+    pub filter_trait_call_args: Option<FilterTraitCallArgsTokens>,
     pub local: bool,
     pub local_cells: bool,
 }
 
 impl NativeFn {
     pub fn ty(&self) -> Type {
-        parse_quote! { turbo_tasks::NativeFunction }
+        parse_quote! { turbo_tasks::macro_helpers::NativeFunction }
     }
 
     pub fn definition(&self) -> TokenStream {
@@ -1065,27 +1113,53 @@ impl NativeFn {
             function_path_string,
             function_path,
             is_method,
+            filter_trait_call_args,
             local,
             local_cells,
         } = self;
 
-        let constructor = if *is_method {
-            quote! { new_method }
+        if *is_method {
+            let arg_filter = if let Some(filter) = filter_trait_call_args {
+                let FilterTraitCallArgsTokens {
+                    filter_owned,
+                    filter_and_resolve,
+                } = filter;
+                quote! {
+                    ::std::option::Option::Some((
+                        #filter_owned,
+                        #filter_and_resolve,
+                    ))
+                }
+            } else {
+                quote! { ::std::option::Option::None }
+            };
+            quote! {
+                {
+                    #[allow(deprecated)]
+                    turbo_tasks::macro_helpers::NativeFunction::new_method(
+                        #function_path_string.to_owned(),
+                        turbo_tasks::macro_helpers::FunctionMeta {
+                            local: #local,
+                            local_cells: #local_cells,
+                        },
+                        #arg_filter,
+                        #function_path,
+                    )
+                }
+            }
         } else {
-            quote! { new_function }
-        };
-
-        quote! {
-            {
-                #[allow(deprecated)]
-                turbo_tasks::NativeFunction::#constructor(
-                    #function_path_string.to_owned(),
-                    turbo_tasks::FunctionMeta {
-                        local: #local,
-                        local_cells: #local_cells,
-                    },
-                    #function_path,
-                )
+            quote! {
+                {
+                    #[allow(deprecated)]
+                    turbo_tasks::macro_helpers::NativeFunction::new_function(
+                        #function_path_string.to_owned(),
+                        turbo_tasks::macro_helpers::FunctionMeta {
+                            local: #local,
+                            local_cells: #local_cells,
+                        },
+                        #function_path,
+                    )
+                }
             }
         }
     }