--- a/turbopack/crates/turbo-tasks-testing/tests/collectibles.rs
+++ b/turbopack/crates/turbo-tasks-testing/tests/collectibles.rs
@@ -192,7 +192,8 @@ async fn my_multi_emitting_function() -> Result<Vc<Thing>> {
 }
 
 #[turbo_tasks::function(operation)]
-async fn my_transitive_emitting_function(key: RcStr, _key2: RcStr) -> Result<Vc<Thing>> {
+async fn my_transitive_emitting_function(key: RcStr, key2: RcStr) -> Result<Vc<Thing>> {
+    let _ = key2;
     my_emitting_function(key).await?;
     Ok(Thing::cell(Thing(0)))
 }
@@ -219,8 +220,9 @@ async fn my_transitive_emitting_function_collectibles(
 async fn my_transitive_emitting_function_with_child_scope(
     key: RcStr,
     key2: RcStr,
-    _key3: RcStr,
+    key3: RcStr,
 ) -> Result<Vc<Thing>> {
+    let _ = key3;
     let thing_op = my_transitive_emitting_function(key, key2);
     let thing_vc = thing_op.connect();
     thing_vc.await?;
@@ -230,7 +232,8 @@ async fn my_transitive_emitting_function_with_child_scope(
 }
 
 #[turbo_tasks::function]
-async fn my_emitting_function(_key: RcStr) -> Result<()> {
+async fn my_emitting_function(key: RcStr) -> Result<()> {
+    let _ = key;
     sleep(Duration::from_millis(100)).await;
     emit(ResolvedVc::upcast::<Box<dyn ValueToString>>(Thing::new(
         123,