@ijjk 2025-01-24
## Failing test suites <!--33463050956-->
Commit: 4dcf00ff6080d991a00de8a55bfcb86b88c901b4

`pnpm test test/integration/404-page-custom-error/test/index.test.js` (turbopack)
- Default 404 Page with custom _error > production mode > should build successfully
- Default 404 Page with custom _error > production mode > should respond to 404 correctly
- Default 404 Page with custom _error > production mode > should render error correctly
- Default 404 Page with custom _error > production mode > should render index page normal
- Default 404 Page with custom _error > production mode > should set pages404 in routes-manifest correctly
- Default 404 Page with custom _error > production mode > should have output 404.html

<details>
<summary>Expand output</summary>

  ● Default 404 Page with custom _error › production mode › should build successfully

    expect(received).toBe(expected) // Object.is equality

    Expected: 0
    Received: 1

      63 |         })
      64 |
    > 65 |         expect(code).toBe(0)
         |                      ^
      66 |
      67 |         appPort = await findPort()
      68 |

      at Object.toBe (integration/404-page-custom-error/test/index.test.js:65:22)

  ● Default 404 Page with custom _error › production mode › should respond to 404 correctly

    TypeError: Invalid URL

      179 | ): Promise<Response> {
      180 |   const url = query ? withQuery(pathname, query) : pathname
    > 181 |   return fetch(getFullUrl(appPort, url), opts)
          |               ^
      182 | }
      183 |
      184 | export function renderViaHTTP(

      at parseURL (../node_modules/.pnpm/node-fetch@2.6.7_encoding@0.1.13/node_modules/node-fetch/lib/index.js:1164:12)
      at new Request (../node_modules/.pnpm/node-fetch@2.6.7_encoding@0.1.13/node_modules/node-fetch/lib/index.js:1210:17)
      at ../node_modules/.pnpm/node-fetch@2.6.7_encoding@0.1.13/node_modules/node-fetch/lib/index.js:1439:19
      at fetch (../node_modules/.pnpm/node-fetch@2.6.7_encoding@0.1.13/node_modules/node-fetch/lib/index.js:1437:9)
      at fetchViaHTTP (lib/next-test-utils.ts:181:15)
      at Object.<anonymous> (integration/404-page-custom-error/test/index.test.js:25:35)

  ● Default 404 Page with custom _error › production mode › should render error correctly

    TypeError: Invalid URL

      179 | ): Promise<Response> {
      180 |   const url = query ? withQuery(pathname, query) : pathname
    > 181 |   return fetch(getFullUrl(appPort, url), opts)
          |               ^
      182 | }
      183 |
      184 | export function renderViaHTTP(

      at parseURL (../node_modules/.pnpm/node-fetch@2.6.7_encoding@0.1.13/node_modules/node-fetch/lib/index.js:1164:12)
      at new Request (../node_modules/.pnpm/node-fetch@2.6.7_encoding@0.1.13/node_modules/node-fetch/lib/index.js:1210:17)
      at ../node_modules/.pnpm/node-fetch@2.6.7_encoding@0.1.13/node_modules/node-fetch/lib/index.js:1439:19
      at fetch (../node_modules/.pnpm/node-fetch@2.6.7_encoding@0.1.13/node_modules/node-fetch/lib/index.js:1437:9)
      at fetchViaHTTP (lib/next-test-utils.ts:181:15)
      at fetchViaHTTP (lib/next-test-utils.ts:190:10)
      at Object.<anonymous> (integration/404-page-custom-error/test/index.test.js:31:37)

  ● Default 404 Page with custom _error › production mode › should render index page normal

    TypeError: Invalid URL

      179 | ): Promise<Response> {
      180 |   const url = query ? withQuery(pathname, query) : pathname
    > 181 |   return fetch(getFullUrl(appPort, url), opts)
          |               ^
      182 | }
      183 |
      184 | export function renderViaHTTP(

      at parseURL (../node_modules/.pnpm/node-fetch@2.6.7_encoding@0.1.13/node_modules/node-fetch/lib/index.js:1164:12)
      at new Request (../node_modules/.pnpm/node-fetch@2.6.7_encoding@0.1.13/node_modules/node-fetch/lib/index.js:1210:17)
      at ../node_modules/.pnpm/node-fetch@2.6.7_encoding@0.1.13/node_modules/node-fetch/lib/index.js:1439:19
      at fetch (../node_modules/.pnpm/node-fetch@2.6.7_encoding@0.1.13/node_modules/node-fetch/lib/index.js:1437:9)
      at fetchViaHTTP (lib/next-test-utils.ts:181:15)
      at fetchViaHTTP (lib/next-test-utils.ts:190:10)
      at Object.<anonymous> (integration/404-page-custom-error/test/index.test.js:36:37)

  ● Default 404 Page with custom _error › production mode › should set pages404 in routes-manifest correctly

    ENOENT: no such file or directory, open '/root/actions-runner/_work/next.js/next.js/test/integration/404-page-custom-error/.next/routes-manifest.json'



  ● Default 404 Page with custom _error › production mode › should have output 404.html

    ENOENT: no such file or directory, open '/root/actions-runner/_work/next.js/next.js/test/integration/404-page-custom-error/.next/server/pages-manifest.json'

      1076 |
      1077 | function readJson(path: string) {
    > 1078 |   return JSON.parse(readFileSync(path, 'utf-8'))
           |                                 ^
      1079 | }
      1080 |
      1081 | export function getBuildManifest(dir: string) {

      at readJson (lib/next-test-utils.ts:1078:33)
      at readJson (lib/next-test-utils.ts:1133:10)
      at getPagesManifest (lib/next-test-utils.ts:1143:25)
      at Object.<anonymous> (integration/404-page-custom-error/test/index.test.js:47:48)

</details>


Read more about building and testing Next.js in [contributing.md](https://github.com/vercel/next.js/blob/canary/contributing.md).